working on mistakes
