+++ src/main/java/school/faang/user_service/mapper/SkillCandidateMapper.java
@@ -7,7 +7,7 @@ import school.faang.user_service.dto.skill.SkillCandidateDto;
 import school.faang.user_service.dto.skill.SkillDto;
 import school.faang.user_service.entity.Skill;
 
-@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
+@Mapper(componentModel = "spring", uses = SkillMapper.class, unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface SkillCandidateMapper {
 
     SkillMapper skillMapper = Mappers.getMapper(SkillMapper.class);
