+++ src/test/java/school/faang/user_service/service/SkillServiceTest.java
@@ -181,51 +181,4 @@ public class SkillServiceTest {
         Mockito.verify(skillRepository, Mockito.times(1)).assignSkillToUser(1L, 1L);
         Mockito.verify(skillServiceTest, Mockito.times(1)).addGuarantee(threeOffers);
     }
-
-    @Test
-    public void testMapperSkillToDto() {
-        Skill skill = new Skill();
-        skill.setTitle("title");
-        skill.setId(1L);
-        SkillDto skillDto = skillMapper.toDto(skill);
-        assertThat(skill.getId() == skillDto.getId()).isTrue();
-        assertThat(skill.getTitle().equals(skillDto.getTitle())).isTrue();
-    }
-
-    @Test
-    public void testMapperDtoToSkill() {
-        SkillDto skillDto = new SkillDto(1L, "title");
-        Skill skill = skillMapper.toEntity(skillDto);
-        assertThat(skill.getTitle().equals(skillDto.getTitle())).isTrue();
-        assertThat(skill.getId() == skillDto.getId()).isTrue();
-    }
-
-    @Test
-    public void testMapperSkillToSkillCandidateDto() {
-        Skill skill = new Skill();
-        skill.setId(1L);
-        skill.setTitle("title");
-        SkillCandidateDto expected = new SkillCandidateDto();
-        expected.setSkill(new SkillDto(1L, "title"));
-        expected.setOffersAmount(1L);
-        SkillCandidateDto actual = skillCandidateMapper.skillToSkillCandidateDto(skill);
-        assertThat(actual.getSkill().equals(expected.getSkill())).isTrue();
-        assertThat(actual.getOffersAmount() == expected.getOffersAmount()).isTrue();
-    }
-
-    @Test
-    public void testMapperListOfSkillsToCandidateDto() {
-        Skill skill = new Skill();
-        skill.setId(1L);
-        skill.setTitle("title");
-        SkillCandidateDto skillCandidateDto = skillCandidateMapper.skillToSkillCandidateDto(skill);
-        skillCandidateDto.setOffersAmount(2);
-        List<Skill> skills = new ArrayList<>();
-        skills.add(skill);
-        skills.add(skill);
-        List<SkillCandidateDto> actual = listOfSkillsCandidateMapper.listToSkillCandidateDto(skills);
-        List<SkillCandidateDto> expected = new ArrayList<>();
-        expected.add(skillCandidateDto);
-        Assertions.assertEquals(expected, actual);
-    }
 }
